db update

+++ src/Domain/Entities/CategoryDiscount.cs
@@ -10,7 +10,7 @@ public class CategoryDiscount : IEntity
   public DateTime RegisterDate { get; set; } = DateTime.UtcNow;
   public DateTime? DeleteDate { get; set; }
   public DateTime EndDate { get; set; }
-  public decimal DiscountPercent { get; set; }
+  public float DiscountPercent { get; set; }
   public string CategoryId { get; set; }
 
   public virtual Category Category { get; set; }
